refactor : 참여자의 수와 결과 수 유효성 검증 gameInfo로 이관
